fix7

+++ src/views/ManageMent.vue
@@ -1,33 +1,107 @@
 <template>
   <div class="home">
     <v-row no-gutters>
-      <v-col cols="2" md="2" sm="1">
-        <MeNu></MeNu>
-      </v-col>
-      <v-col cols="10" md="10" sm="11">
-        <v-app-bar style="background-color: #d4d4d4" elevate-on-scroll>
+      <v-col>
+        <v-app-bar             
+        scroll-target="#scrolling-techniques-7"
+        height="100"
+        style="background-color: #cdcdcd"
+        class="hidden-sm-and-down col00">
           <v-container>
             <v-row>
-              <v-span style="margin-top: 5px; font-size: 18px"
-                >จัดการข้อมูล</v-span
-              >
+
+              <p style="margin-top: 30px; font-size: 25px ; color:#ffffff ;" class="hidden-sm-and-down">
+                ✍ จัดการข้อมูล</p>
+            
+              
               <v-spacer></v-spacer>
+              <v-text-field
+              style=" margin-left:500px; margin-top:28px;"
+              class=" mr-5  hidden-sm-and-down"
+              v-model="search"
+              append-icon="mdi-magnify"
+              label="Search"
+              solo
+            ></v-text-field>
               <v-btn
+                class="hidden-sm-and-down"
                 href="/management/add"
                 dark
-                small
+                large
                 fab
-                style="background-color: #00bf00"
+                style="background-color: #00bf00 ; margin-right: 5px ; margin-top: 17px"
                 elevation="5"
               >
-                <v-icon> mdi-package-variant-closed-plus </v-icon>
+                <v-icon large> mdi-package-variant-closed-plus </v-icon>
               </v-btn>
             </v-row>
           </v-container>
         </v-app-bar>
+        <!-- ---------------------------------------sm and down------------------------------------- -->
+        <div>
+          <v-app-bar             
+          scroll-target="#scrolling-techniques-7"
+          height="100"
+          style="background-color: #cdcdcd"
+          class="hidden-md-and-up col00">
+          <v-text-field
+          style=" margin-right:15px ;"
+          class="mt-8 hidden-md-and-up"
+          v-model="search"
+          append-icon="mdi-magnify"
+          label="Search"
+          solo
+        ></v-text-field>
+          <v-btn
+          class="hidden-md-and-up"
+          href="/management/add"
+          dark
+          fab
+          style="background-color: #00bf00 ; margin-right: 5px ;  margin-top: 0px ;"
+          elevation="5"
+          
+        >
+              <v-icon> mdi-package-variant-closed-plus </v-icon>
+            </v-btn>
+          </v-app-bar>
+        </div>
+          <!-- ---------------------------------------sm and down------------------------------------- -->
         <v-row>
           <v-col>
-            <v-simple-table height="835px">
+            <v-data-table 
+              height="76vh"
+              :headers="headers"
+              :items="StocK1"
+              :search="search"
+              :footer-props="footerProps"
+            >
+
+              <template v-slot:[`item.action`]="{ item }" >
+                <td style="width: 180px;">
+                <v-btn 
+                  @click="AlertClick(item.id)"
+                  small
+                  class="ma-2"
+                  width="50"
+                  style="background-color: #ff3232"
+                  dark
+                  elevation="5"
+                >
+                ลบ
+                </v-btn>
+                <v-btn
+                small
+                class="ma-2"
+                width="50"
+                style="background-color: #356bff"
+                dark
+                elevation="5"
+                >แก้ไข</v-btn
+              ></td>
+              </template>
+            </v-data-table>
+            
+            <!-- <v-simple-table height="84.4vh"> 
               <template v-slot:default>
                 <tbody>
                   <tr v-for="item in StocK1" :key="item.id">
@@ -36,8 +110,6 @@
                     <td>{{ item.qty }}</td>
                     <td class="text-right">
                       <v-btn
-                        @click="Testclick"
-                        
                         small
                         class="ma-2"
                         width="50"
@@ -48,7 +120,7 @@
                       >
 
                       <v-btn
-                        @click="Alertclick(item.id)"
+                        @click="AlertClick(item.id)"
                         small
                         class="ma-2"
                         width="50"
@@ -62,7 +134,7 @@
                   </tr>
                 </tbody>
               </template>
-            </v-simple-table>
+            </v-simple-table> -->
           </v-col>
         </v-row>
       </v-col>
@@ -72,16 +144,29 @@
 
 <script>
 import Swal from "sweetalert2";
-import MeNu from "@/components/MeNu.vue";
 export default {
   data() {
     return {
       StocK1: [],
       ID: "",
       number: 0,
+      valll:false,
+      search:'',
+      footerProps: {'items-per-page-options': [11,20,25]},
+      headers: [
+          {
+            text: 'Name',
+            align: 'start',
+            sortable: false,
+            value: 'name',
+          },
+          { text: 'Codenumber', value: 'code' },
+          { text: 'Quantity', value: 'qty' },
+          { text: '', value: 'action' ,},
+        ],
     };
   },
-  components: { MeNu },
+
   mounted() {
     this.axios.get("http://localhost:3000/stock").then((response) => {
       this.StocK1 = response.data;
@@ -94,13 +179,12 @@ export default {
         .delete(`http://localhost:3000/delete`, { data: { id: codeid } })
         .then((response) => {
           console.log(response);
-          window.location.reload();
         })
         .catch(function (error) {
           document.write("ส่งข้อมูลไม่สำเร็จ" + "<br>" + "สาเหตุ :" + error);
         });
     },
-    Alertclick(list) {
+    AlertClick(list) {
       this.number = list;
       let codeid = this.number;
       Swal.fire({
@@ -114,8 +198,14 @@ export default {
         cancelButtonText:"ยกเลิก"
       }).then((result) => {
         if (result.isConfirmed) {
-          Swal.fire("ลบสำเร็จ!", "ข้อมูลใน stock ถูกลบแล้ว", "success");
-          this.DeleteItem(codeid)
+          Swal.fire({title:"ลบสำเร็จ!", text:"ข้อมูลใน stock ถูกลบแล้ว", 
+          icon:"success", showConfirmButton: false,})
+          .then(() => {
+            if(Swal.DismissReason.backdrop){
+              this.DeleteItem(codeid);
+              window.location.reload();
+            }
+          })
         }
       });
     },
@@ -123,5 +213,12 @@ export default {
 };
 </script>
 
-<style>
+<style scoped>
+  .high {
+    height: 500px;
+  }
+  .col00{
+    background: rgb(54,54,54);
+    background: linear-gradient(90deg, rgba(54,54,54,1) 0%, rgba(173,173,173,1) 100%);
+  }
 </style>
